Remove redundant main module
