+++ pyproject.toml
@@ -14,4 +14,4 @@ dependencies = [
 [project.gui-scripts]
 blackjack-sim = "blackjack.__main__:run_gui"
 [project.scripts]
-blackjack-cli = "blackjack.main:run_cli"
+blackjack-cli = "blackjack.__main__:run_cli"
